+++ src/app.js
@@ -3,8 +3,7 @@ import DustService from '@/api/modules/dust';
 import {mainCardTemplate, forecastCardTemplate} from '@/utils/templates';
 import {getPmStatusEmoji} from '@/utils/utils';
 import transformCoord from '@/utils/transformCoord';
-import searchDate from '@/utils/searchDate';
-import findForecastAreaIndex from './utils/findArea';
+import getForecast from './utils/getForecast';
 
 function DustApp() {
   this.init = async () => {
@@ -14,6 +13,7 @@ function DustApp() {
   };
   this.getDustStatus = async position => {
     const {latitude, longitude} = position.coords;
+    // const [latitude, longitude]= [36.4810123,127.2883183]
     const msrstnList = await MapService.getMyAddress({latitude, longitude});
     const address = msrstnList.data.results[0].formatted_address.split(' ');
     const [tmX, tmY] = transformCoord(longitude, latitude);
@@ -22,43 +22,8 @@ function DustApp() {
       longitude: tmX,
     });
 
-    const [year, month, date] = searchDate();
-    const forecastInfo = await DustService.getPMForecast({year, month, date});
-    const totalForecast = forecastInfo.data.forecast;
-    const releaseTime = totalForecast[0].f_data_time;
-    const pickForecast = totalForecast.filter(
-      val => val.f_data_time === releaseTime,
-    );
     const [, province, city] = address;
-
-    const convertGrade = grade => {
-      switch (grade) {
-        case '좋음':
-          return '1';
-        case '보통':
-          return '2';
-        case '나쁨':
-          return '3';
-        case '매우나쁨':
-          return '4';
-        case '매우 나쁨':
-          return '4';
-        default:
-          return '-1';
-      }
-    };
-    const forecastAreaIndex = findForecastAreaIndex(province, city);
-    const finalForecast = pickForecast.map(data => {
-      const makeGradeArray = data.informGrade.split(',');
-      let [area, grade] = makeGradeArray[forecastAreaIndex].split(':');
-      area = area.trim();
-      grade = grade.trim();
-      return {
-        date: data.informData,
-        grade: getPmStatusEmoji(convertGrade(grade)),
-        area,
-      };
-    });
+    const finalForecast = await getForecast(province, city);
     this.render(address, arpltnInfo.data.pm10, finalForecast);
   };
   this.render = (address, pm10, finalForecast) => {
